+++ tests/test_install.py
@@ -1,0 +1,4 @@
+#Tests the installation from the PyPI
+def test_instantiate_form():
+    import delphivcl
+    frm = delphivcl.Form
